split user auth
