+++ back/src/app.module.ts
@@ -8,14 +8,13 @@ import { typeormconfig } from './config/typeorm.config';
 import { MulterModule } from '@nestjs/platform-express';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
-import { ChatModule } from './chat/chat.module';
-import { ChatModule } from './chat/chat.module';
+import { AuthenticationModule } from './authentication/authentication.module';
 
 @Module({
   imports: [
     UserModule,
     TypeOrmModule.forRoot(typeormconfig),
-    ChatModule,
+    AuthenticationModule,
     // CorsModule.forRoot({ origin: '*' })
     // MulterModule.register({dest: './files'}),
     //  ServeStaticModule.forRoot({rootPath: join(__dirname, '..', 'files')})
