+++ back/src/authentication/authentication.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { AuthenticationController } from './authentication.controller';
+import { AuthenticationService } from './authentication.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from 'src/user/user.entity';
+
+
+@Module({
+  imports: [TypeOrmModule.forFeature([User])],
+  controllers: [AuthenticationController],
+  providers: [AuthenticationService]
+})
+export class AuthenticationModule { }
