+++ back/src/user/user.module.ts
@@ -4,7 +4,7 @@ import { UserService } from './user.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { JwtModule } from '@nestjs/jwt'
-import { JwtAuthGuard } from './auth.guard';
+import { JwtAuthGuard } from './user.guard';
 import { PassportModule } from '@nestjs/passport';
 import { JwtStrategy } from './user.strategy';
